The conditional expression operator '?' apparently gets its expression type from second expression, which on line 240 was an unsigned int.  This caused the -1 to be converted to an unsigned int, which gives an error because the return type of getfs () is int.  Cast full_scale to an (int) so that the -1 is not implicitly cast to an (unsigned int).

+++ Display/dispwind.h
@@ -236,8 +236,9 @@ class win_attributed :
                                                      manual     = TRUE;
                                                    }
                                 int manuallyscaled() { return manual; }
-                                int getfs() { return (manual ? full_scale :
-                                                              -1); }
+                                int getfs() { return (manual 
+						      ? (int) full_scale
+						      : -1); }
                                 unsigned int getfsval() { return full_scale; }
                                 void setfsval(unsigned int fs) { full_scale = fs; }
                                 // Get/set counts axis attribute
